pivot_drive_controller bug fixes - now builds without errors

+++ src/ros/rover/controllers/pivot_drive_controller/src/speed_limiter.cpp
@@ -24,9 +24,9 @@
 namespace pivot_drive_controller
 {
 SpeedLimiter::SpeedLimiter(
-  bool has_velocity_limits, bool has_acceleration_limits, bool has_jerk_limits, double min_velocity,
-  double max_velocity, double min_acceleration, double max_acceleration, double min_jerk,
-  double max_jerk)
+  bool has_velocity_limits, bool has_acceleration_limits, bool has_jerk_limits, float min_velocity,
+  float max_velocity, float min_acceleration, float max_acceleration, float min_jerk,
+  float max_jerk)
 : has_velocity_limits_(has_velocity_limits),
   has_acceleration_limits_(has_acceleration_limits),
   has_jerk_limits_(has_jerk_limits),
@@ -74,9 +74,9 @@ SpeedLimiter::SpeedLimiter(
   }
 }
 
-double SpeedLimiter::limit(double & v, double v0, double v1, double dt)
+float SpeedLimiter::limit(float & v, float v0, float v1, float dt)
 {
-  const double tmp = v;
+  const float tmp = v;
 
   limit_jerk(v, v0, v1, dt);
   limit_acceleration(v, v0, dt);
@@ -85,9 +85,9 @@ double SpeedLimiter::limit(double & v, double v0, double v1, double dt)
   return tmp != 0.0 ? v / tmp : 1.0;
 }
 
-double SpeedLimiter::limit_velocity(double & v)
+float SpeedLimiter::limit_velocity(float & v)
 {
-  const double tmp = v;
+  const float tmp = v;
 
   if (has_velocity_limits_)
   {
@@ -97,16 +97,16 @@ double SpeedLimiter::limit_velocity(double & v)
   return tmp != 0.0 ? v / tmp : 1.0;
 }
 
-double SpeedLimiter::limit_acceleration(double & v, double v0, double dt)
+float SpeedLimiter::limit_acceleration(float & v, float v0, float dt)
 {
-  const double tmp = v;
+  const float tmp = v;
 
   if (has_acceleration_limits_)
   {
-    const double dv_min = min_acceleration_ * dt;
-    const double dv_max = max_acceleration_ * dt;
+    const float dv_min = min_acceleration_ * dt;
+    const float dv_max = max_acceleration_ * dt;
 
-    const double dv = std::clamp(v - v0, dv_min, dv_max);
+    const float dv = std::clamp(v - v0, dv_min, dv_max);
 
     v = v0 + dv;
   }
@@ -114,21 +114,21 @@ double SpeedLimiter::limit_acceleration(double & v, double v0, double dt)
   return tmp != 0.0 ? v / tmp : 1.0;
 }
 
-double SpeedLimiter::limit_jerk(double & v, double v0, double v1, double dt)
+float SpeedLimiter::limit_jerk(float & v, float v0, float v1, float dt)
 {
-  const double tmp = v;
+  const float tmp = v;
 
   if (has_jerk_limits_)
   {
-    const double dv = v - v0;
-    const double dv0 = v0 - v1;
+    const float dv = v - v0;
+    const float dv0 = v0 - v1;
 
-    const double dt2 = 2. * dt * dt;
+    const float dt2 = 2. * dt * dt;
 
-    const double da_min = min_jerk_ * dt2;
-    const double da_max = max_jerk_ * dt2;
+    const float da_min = min_jerk_ * dt2;
+    const float da_max = max_jerk_ * dt2;
 
-    const double da = std::clamp(dv - dv0, da_min, da_max);
+    const float da = std::clamp(dv - dv0, da_min, da_max);
 
     v = v0 + dv0 + da;
   }
